feat: 主动安全防护系统附件服务器

+++ active-safety/src/main/java/com/avenger/jt808/FileData.java
@@ -9,6 +9,6 @@ import lombok.Data;
 @Data
 public class FileData {
     private String fileName;
-    private int offset;
+    private long offset;
     private byte[] data;
 }
